tests: add invalid config test coverage
